Moved user endpoints in to own file
