+++ P7Internet.Frontend/react-app/src/components/LoginBox.jsx
@@ -3,7 +3,7 @@ import cross from "../data/cross.svg";
 import { useDispatch, useSelector } from "react-redux";
 import { userActions } from "../features/userSlice";
 import { toast } from 'react-toastify';
-import { useUserCreateMutation, useUserLoginMutation } from "../services/userApiSlice";
+import { useUserCreateMutation, useUserLoginMutation } from "../services/usersEndpoints";
 import 'react-toastify/dist/ReactToastify.css';
